+++ src/Component/OpenApi3/Tests/fixtures/github/expected/Endpoint/ProjectsCreateCard.php
@@ -29,7 +29,7 @@ class ProjectsCreateCard extends \Github\Runtime\Client\BaseEndpoint implements 
     public function getBody(\Symfony\Component\Serializer\SerializerInterface $serializer, $streamFactory = null): array
     {
         if (isset($this->body)) {
-            return [['Content-Type' => ['application/json']], json_encode($this->body)];
+            return [['Content-Type' => ['application/json']], $serializer->serialize($this->body, 'json')];
         }
         return [[], null];
     }