+++ src/Component/OpenApi3/Tests/fixtures/github/expected/Endpoint/UsersAddEmailForAuthenticated.php
@@ -24,7 +24,7 @@ class UsersAddEmailForAuthenticated extends \Github\Runtime\Client\BaseEndpoint 
     public function getBody(\Symfony\Component\Serializer\SerializerInterface $serializer, $streamFactory = null): array
     {
         if (isset($this->body)) {
-            return [['Content-Type' => ['application/json']], json_encode($this->body)];
+            return [['Content-Type' => ['application/json']], $serializer->serialize($this->body, 'json')];
         }
         return [[], null];
     }